% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSignal.R
\name{advanced_filter}
\alias{advanced_filter}
\title{Advanced Signal Quality Evaluation}
\usage{
advanced_filter(signal, pressure_start_point = 3L, silent = TRUE)
}
\arguments{
\item{signal}{A numeric vector representing the signal.}

\item{pressure_start_point}{(Optional) The starting point of the pressure in
the signal. Defaults to 3L.}

\item{silent}{(Optional) A logical for running the function silently
(without printouts). Defaults to TRUE.}
}
\value{
A list containing the calculated score and detailed results of each
 penalization.
}
\description{
Evaluates the quality of a provided signal using advanced criteria,
and assigns a score based on the outcome of these checks.
}
\details{
The function performs advanced checks on the signal and penalizes specific
 characteristics:
1. Stabilization Slope Check: If the stabilization phase slope
 (multiplied by 100) is greater than 0.0001, the score is penalized by this
 value.
2. Drop vs. Peak Increment Check: If the drop before stabilization exceeds
 45% of the signal's peak increment, the score is penalized by ten times the
 excess drop.
3. Stabilization Level Check: If the stabilization level is lower than the
 peak minimum, the score is penalized based on the distance between these
 points (multiplied by 10).
}
\examples{
sample_signal <- rnorm(100)
start_point <- 5
filter_results <- advanced_filter(sample_signal, start_point)

}
