% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalHelpers.R
\name{calculate_discrete_derivatives}
\alias{calculate_discrete_derivatives}
\title{Calculate the Discrete First and Second Derivatives of a Numeric Vector}
\usage{
calculate_discrete_derivatives(vec)
}
\arguments{
\item{vec}{A numeric vector representing a discrete function over an ordered domain.}
}
\value{
A list containing the discrete first and second derivatives of the vector.
The first element, 'first', contains the differences representing the first derivative.
The second element, 'second', contains the differences of the first derivative,
representing the second derivative.
}
\description{
This function computes the discrete first and second derivatives of a numeric vector.
The discrete first derivative is defined as the difference between consecutive points,
providing a measure of the instantaneous rate of change. The discrete second derivative
is the difference of the first derivative values, offering a measure of the acceleration
or the rate at which the rate of change itself is changing, indicative of the vector's
concavity or convexity.
}
\examples{
vec <- c(1, 2, 4, 7, 11)
derivatives <- calculate_discrete_derivatives(vec)
# `derivatives$first` contains the discrete first derivative of `vec`.
# `derivatives$second` contains the discrete second derivative of `vec`.

}
