% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{configure_data_env}
\alias{configure_data_env}
\title{Configure Data Environment for PSO Modeling}
\usage{
configure_data_env(
  config,
  data,
  excluded_cols = NULL,
  signal_names,
  predictors_names,
  vsvr_response,
  extra_col_name = NULL,
  initial_prediction_values = c(1),
  multi = FALSE
)
}
\arguments{
\item{config}{A list of configuration options for the brain modeler.}

\item{data}{A data frame containing the dataset to be optimized.}

\item{excluded_cols}{(Optional) Vector of column names to be excluded from
the optimization. Defaults to NULL.}

\item{signal_names}{Vector of names representing signal columns.}

\item{predictors_names}{Vector of predictor names for the model.}

\item{vsvr_response}{Name of the column representing the VSVR response.}

\item{extra_col_name}{(Optional) Additional column name to be included. Only
used if `multi` is TRUE. Defaults to NULL.}

\item{initial_prediction_values}{(Optional) Initial prediction values for the
model. Defaults to c(1).}

\item{multi}{(Optional) Logical indicating if the model should be multivariate.
If TRUE, additional processing is performed for the extra column. Defaults
to FALSE.}
}
\value{
A list containing the following elements:
  - `pressure_df`: Data frame with pressure step added.
  - `processed_data`: The processed data frame.
  - `data_partitions`: Data partitions created by blocked cross-validation.
  - `NORM_SIGNAL_NAMES`: Normalized signal names.
  - `NORM_PREDICTORS_NAMES`: Normalized predictor names.
  - `NORM_VSVR_RESPONSE`: Normalized VSVR response name.
  - `VSVR_TOL`: VSVR tolerance from the configuration.
  - `INITIAL_PREDICTION_VALUES`: Initial prediction values,
      adjusted if `multi` is TRUE.
  - `svm_cache_size`: SVM cache size from the configuration.
}
\description{
This function prepares the data environment for PSO modeling.
It processes the input data and sets up the necessary components and
configurations based on the provided parameters. It handles the addition of
a pressure step, processing of the data frame, and blocked cross-validation
partitioning. It also normalizes signal names and handles additional
configurations for multivariate models.
}
\examples{
\dontrun{
  configure_data_env(...)
}

}
