% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{configure_pso_brain_modeler}
\alias{configure_pso_brain_modeler}
\title{Configure the Brain Modeler for PSO}
\usage{
configure_pso_brain_modeler(
  seed = 123,
  bcv_folds = 5,
  bcv_validation_size = 0.2,
  pressure_signal_start = 3,
  pressure_signal_response_size = 40,
  butter_filter_order = 2,
  butter_filter_fs = 0.2,
  max_lag_number = 8,
  vsvr_tolerance = 1
)
}
\arguments{
\item{seed}{Seed for random number generation. Default: 123.}

\item{bcv_folds}{Number of folds for blocked cross-validation. Default: 5.}

\item{bcv_validation_size}{Proportion of data for validation in blocked CV. Default: 0.2.}

\item{pressure_signal_start}{Starting point for pressure signal. Default: 3.}

\item{pressure_signal_response_size}{Size of the response signal for pressure. Default: 40.}

\item{butter_filter_order}{Order of the Butterworth filter. Default: 2.}

\item{butter_filter_fs}{Sampling frequency for the Butterworth filter. Default: 0.2.}

\item{max_lag_number}{Maximum number of lags to consider. Default: 8.}

\item{vsvr_tolerance}{Tolerance level for VSVR. Default: 1.}
}
\value{
A list containing the configuration with class "PSOBrainModelerConfig".
}
\description{
This function sets the configuration for the PSO brain modeler.
}
