% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{configure_psoptim_control}
\alias{configure_psoptim_control}
\title{Configure the control parameters for psoptim}
\usage{
configure_psoptim_control(
  pso_swarm_size = 5,
  pso_max_iterations = 100,
  pso_max_fn_calls = 200,
  pso_informed_swarm = 0.5,
  pso_global_exp_const = 10,
  pso_maxit_without_improvement = 50,
  pso_restart_tolerance = 0.5,
  pso_hybrid_type = "improved",
  pso_type = "SPSO2011",
  pso_vectorization = TRUE,
  pso_hybrid_control = list(maxit = 1)
)
}
\arguments{
\item{pso_swarm_size}{Size of the particle swarm. Default: 5.}

\item{pso_max_iterations}{Maximum number of iterations for PSO. Default: 100.}

\item{pso_max_fn_calls}{Maximum number of function calls for PSO. Default: 200.}

\item{pso_informed_swarm}{Proportion of swarm to inform. Default: 0.5.}

\item{pso_global_exp_const}{Global expansion constant for PSO. Default: 10.}

\item{pso_maxit_without_improvement}{Maximum iterations without improvement. Default: 50.}

\item{pso_restart_tolerance}{Tolerance for restarting PSO. Default: 0.5.}

\item{pso_hybrid_type}{Type of hybrid method for PSO. Default: "improved".}

\item{pso_type}{Type of PSO method. Default: "SPSO2011".}

\item{pso_vectorization}{Logical indicating if PSO should be vectorized. Default: TRUE.}

\item{pso_hybrid_control}{List of control parameters for the hybrid method.}
}
\value{
A list containing the psoptim configuration with class "PSOBrainModelerPSOPTIMConfig".
}
\description{
Sets the configuration for the psoptim optimizer.
}
