% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{cross_validate_partition_helper}
\alias{cross_validate_partition_helper}
\title{Helper function for cross-validation of partitions}
\usage{
cross_validate_partition_helper(
  cost,
  nu,
  gamma,
  col_lags,
  data_list,
  silent = FALSE
)
}
\arguments{
\item{cost}{A numeric value for the cost parameter.}

\item{nu}{A numeric value for the nu parameter.}

\item{gamma}{A numeric value for the gamma parameter.}

\item{col_lags}{Numeric vector indicating the lags for the columns.}

\item{data_list}{A list containing various data configurations}

\item{silent}{A logical for run the function silently (without printouts). Defaults to FALSE.}
}
\value{
Result from the cross_validate_partition function.
}
\description{
This function constructs parameters for the cross_validate_partition function.
It sets the configuration related to the SVR model, data partitions, and normalization.
}
