% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{display_params_message}
\alias{display_params_message}
\title{Display Formatted Message Based on Parameters List}
\usage{
display_params_message(params_list)
}
\arguments{
\item{params_list}{A list containing parameters including cost, nu, gamma,
column lags, and response lags.}
}
\description{
This function constructs and displays a message detailing the values of the
parameters in the provided list. It specifically formats and includes
information about cost, nu, gamma (if not NULL), and lags.
}
\examples{
params_list <- list(cost = 1.23, nu = 0.45, gamma = NULL, col_lags = 1:3,
 response_lags = 4:5)
display_params_message(params_list)

}
