% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{display_pso_message}
\alias{display_pso_message}
\title{Display PSO Message}
\usage{
display_pso_message(cor, err, basic, advanced)
}
\arguments{
\item{cor}{A numeric value or vector representing the average correlation.
This parameter should provide the average correlation scores
calculated
by your process or model.}

\item{err}{A numeric value or vector representing the average mean squared
error (MSE).
This parameter should contain the average MSE values calculated
by your process or model.}

\item{basic}{A numeric or character value representing the basic filter
signal score.
This score or flag indicates the result of a basic filtering
process or evaluation.}

\item{advanced}{A numeric or character value representing the advanced filter
signal score.
This score or flag indicates the result of an advanced
filtering process or evaluation.}
}
\value{
This function does not return a value. It prints the input parameters
 in a formatted manner to the R console.
}
\description{
This function displays a message with various statistical measures and
signal filter scores.
It is designed to output the average correlation (AVG COR), average mean
squared error (AVG MSE), basic filter signal score, and advanced filter
signal score.
}
\examples{
display_pso_message(0.85, 0.02, "Pass", "High")

}
