% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Load Initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the library.}

\item{pkgname}{The name of the package.}
}
\description{
Initializes the global variables stored in the .psoBrainModelerEnv environment
when the package is loaded.
}
\keyword{internal}
