% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\docType{data}
\name{.psoBrainModelerEnv}
\alias{.psoBrainModelerEnv}
\title{Internal PSO Brain Modeler Environment}
\format{
An object of class \code{environment} of length 7.
}
\usage{
.psoBrainModelerEnv
}
\description{
An environment to store global variables related to the PSO Brain Modeler.
}
\keyword{datasets}
