% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSignal.R
\name{evaluate_signal_quality}
\alias{evaluate_signal_quality}
\title{Signal Quality Evaluation}
\usage{
evaluate_signal_quality(signal, pressure_start_point = 3L, silent = FALSE)
}
\arguments{
\item{signal}{A numeric vector representing the signal.}

\item{pressure_start_point}{(Optional) The starting point of the pressure. Defaults to 3L.}

\item{silent}{(Optional) A logical for run the function silently (without printouts). Defaults to FALSE.}
}
\value{
A numeric score indicating the quality of the signal.
}
\description{
Evaluates the quality of a provided signal based on specific criteria.
}
\details{
The function applies both basic and advanced filters to evaluate the signal's quality.
The basic filter checks for conditions related to peaks, minimums, variance, and range.
The advanced filter performs checks on stabilization and drop rates.
}
\examples{
 sample_signal <- rnorm(40)
 start_point <- 5
 quality_score <- evaluate_signal_quality(sample_signal, start_point)

}
