% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{exclude_signals_dataframe}
\alias{exclude_signals_dataframe}
\title{Exclude Columns in a Data Frame by Names}
\usage{
exclude_signals_dataframe(df, signal_names = NULL)
}
\arguments{
\item{df}{A data frame to filter.}

\item{signal_names}{A character vector specifying the names of the columns to exclude from the output.}
}
\value{
A data frame without the specified columns. Returns NULL if any of the specified columns are not found.
}
\description{
This function returns a subset of the given data frame excluding the specified columns.
It validates if the specified column names exist in the data frame before exclusion.
}
