% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{extract_params_list}
\alias{extract_params_list}
\title{Extract and Format Parameters List for PSO Modeling}
\usage{
extract_params_list(params, model_parameters, multi = FALSE)
}
\arguments{
\item{params}{A vector of parameters for PSO modeling.}

\item{model_parameters}{A list of model parameters, including valid lengths
and number of lags.}

\item{multi}{Logical indicating if the model is multivariate. Defaults to
FALSE.}
}
\value{
A list of formatted parameters, including cost, nu,
 gamma (if applicable), response lags, and column lags.
}
\description{
This function extracts parameters necessary for PSO modeling from a given
list, adjusts for the presence of a gamma parameter, and formats them
accordingly. It also handles adjustments for multivariate models.
}
\examples{
params <- c(1, 2, 3, 4, 5)
model_parameters <- list(valid_lengths = c(4, 5), n_lags = 2)
extract_params_list(params, model_parameters)

}
