% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockedCrossValidation.R
\name{generate_and_evaluate_response_cv}
\alias{generate_and_evaluate_response_cv}
\title{Generate and Evaluate Response for Cross-Validation}
\usage{
generate_and_evaluate_response_cv(params)
}
\arguments{
\item{params}{A list containing all necessary parameters for generating and
 evaluating the response.
The list should include:
\itemize{
  \item \code{data}: The training data as a dataframe.
  \item \code{col_lags}: Numeric or integer vector specifying the lags for
   each column.
  \item \code{response_lags}: Numeric or integer for the response variable
   lags.
  \item \code{initial_column_values}: Initial values for each column as a
   numeric vector or list.
  \item \code{prediction_initial_value}: Initial value for prediction.
  \item \code{cost}: The cost parameter for the SVR model.
  \item \code{nu}: The nu parameter for the SVR model.
  \item \code{gamma}: The gamma parameter for the SVR model.
  \item \code{data_list}: Additional data structures or context for the
   response prediction.
  \item \code{included_model}: The trained SVR model.
  \item \code{silent}: Logical flag to run the function silently.
}}
}
\value{
A list containing the generated response signal, associated warnings,
 support vectors count, basic filter score, and advanced score along with
 its results.
}
\description{
This function generates response predictions, evaluates signal quality, and
applies an advanced filter as part of the cross-validation process in an SVR
model. It is specifically designed to be used within the
`cross_validate_partition` function.
}
\examples{
\dontrun{
  generate_and_evaluate_response_cv(...)
}

}
