% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{generate_signal_response_predictions_helper}
\alias{generate_signal_response_predictions_helper}
\title{Helper Function to Generate Signal Response Predictions}
\usage{
generate_signal_response_predictions_helper(
  data,
  col_lags,
  response_lags,
  initial_column_values,
  prediction_initial_value,
  cost,
  nu,
  gamma,
  data_list,
  included_model = NULL
)
}
\arguments{
\item{data}{Data frame containing the dataset to be processed.}

\item{col_lags}{Numeric vector indicating the lags for the columns.}

\item{response_lags}{Numeric vector indicating the response lags. Defaults
to NULL.}

\item{initial_column_values}{Numeric vector indicating initial values for
columns.}

\item{prediction_initial_value}{Numeric value indicating the initial
prediction value.}

\item{cost}{Numeric value for the cost parameter in the SVR model.}

\item{nu}{Numeric value for the nu parameter in the SVR model.}

\item{gamma}{Numeric value for the gamma parameter in the SVR model.
Defaults to NULL.}

\item{data_list}{A list containing various data configurations and settings
relevant to the SVR model and dataset.}

\item{included_model}{(Optional) The pre-trained SVR model to be used.
If NULL, a new model is trained.}
}
\value{
The result from the generate_signal_response_predictions function,
 typically predictions of signal responses alongside any relevant model
 parameters and statistics.
}
\description{
This function prepares parameters for the generate_signal_response_predictions
function.
It aims to predict signal responses based on the SVR model and the provided
dataset, taking into account various parameters such as lags, initial values,
and SVR model settings.
}
\examples{
\dontrun{
  generate_signal_response_predictions_helper(...)
}

}
