% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{get_model_parameters}
\alias{get_model_parameters}
\title{Auxiliary Function to Get Model Parameters}
\usage{
get_model_parameters(model, multi = FALSE)
}
\arguments{
\item{model}{Character string specifying the type of model.
Should be one of "FIR", "NFIR", "ARX", or "NARX".}

\item{multi}{(Optional) Logical indicating if the model is multivariate.
Defaults to FALSE.}
}
\value{
A list containing two elements:
  - `valid_lengths`: Vector of valid parameter lengths for the specified model.
  - `n_lags`: Number of lags to be used for the model.
}
\description{
This function retrieves the valid parameter lengths and number of lags for
 different model types.
It supports both univariate and multivariate cases for FIR, NFIR, ARX, and
 NARX models.
}
\examples{
get_model_parameters("FIR", TRUE)
get_model_parameters("ARX", FALSE)

}
