% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchSVR.R
\name{grid_eval}
\alias{grid_eval}
\title{Evaluate a grid of parameters}
\usage{
grid_eval(params, progress_bar = NULL)
}
\arguments{
\item{params}{A list containing the parameters for the grid evaluation.}

\item{progress_bar}{A progress bar object for tracking progress. Optional.}
}
\description{
This function evaluates a grid of parameters for signal processing. It
iterates over response lags,
displays a message (if not silent), and calls a function to evaluate the
signal on the grid.
It supports a progress bar update after each iteration.
}
\examples{
\dontrun{
  grid_eval(params, progress_bar)
}

}
