% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchSVR.R
\name{grid_search_main_loop}
\alias{grid_search_main_loop}
\title{Grid Search Main Loop}
\usage{
grid_search_main_loop(
  data_env,
  processed_data,
  data_partitions,
  bcv_folds = 5,
  lags_column = c(8),
  initial_column_values = c(1),
  lags_response = NULL,
  initial_prediction_value = 1,
  model_params,
  nu,
  cost,
  training_list_name = "training",
  validation_list_name = "validation",
  is_silent_mode = TRUE,
  should_plot_response = FALSE,
  progress_bar = NULL,
  generate_response_predictions_cv = FALSE,
  basic_filter_check = TRUE,
  show_progress_bar = FALSE
)
}
\arguments{
\item{data_env}{Data environment object containing datasets and parameters
for the grid search.}

\item{processed_data}{Dataframe of processed data set to be used in the model.}

\item{data_partitions}{Data partitions object for cross-validation.}

\item{bcv_folds}{Integer specifying the number of folds for cross-validation,
defaults to 5.}

\item{lags_column}{Integer or vector specifying the number of lags for the
columns, defaults to c(8).}

\item{initial_column_values}{Numeric or vector providing initial values for
columns, defaults to c(1).}

\item{lags_response}{Integer or vector specifying the number of lags for the
response, defaults to NULL.}

\item{initial_prediction_value}{Numeric value indicating the initial value
for prediction, defaults to 1.}

\item{model_params}{List of model parameters including kernel type.}

\item{nu}{Numeric value representing the nu parameter for the SVR model.}

\item{cost}{Numeric value representing the cost parameter for the SVR model.}

\item{training_list_name}{Character string naming the list of training data,
defaults to 'training'.}

\item{validation_list_name}{Character string naming the list of validation
data, defaults to 'validation'.}

\item{is_silent_mode}{Logical flag indicating if the function should operate
in silent mode, suppressing messages and output, defaults to TRUE.}

\item{should_plot_response}{Logical flag indicating whether the response
variable should be plotted, defaults to FALSE.}

\item{progress_bar}{Progress bar object configured for the grid search
process.}

\item{generate_response_predictions_cv}{Logical flag indicating if response
predictions should be generated, defaults to FALSE.}

\item{basic_filter_check}{Logical flag indicating if basic filters should
be applied, defaults to TRUE.}

\item{show_progress_bar}{Disables progress bar. Defaults to FALSE.}
}
\value{
A list containing the results of each iteration of the grid search,
 including model performance metrics and potentially the generated response
 predictions.
}
\description{
This function executes the main loop of the grid search process for Support
Vector Regression.
It iterates over different combinations of cost and nu parameters, evaluating
the model's performance for each combination.
}
\examples{
\dontrun{
# Assuming 'data_env', 'processed_data', and 'model_params' are predefined
results <- grid_search_main_loop(
  data_env = data_env,
  processed_data = processed_data,
  data_partitions = data_env$data_partitions, ...)
}
}
