% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{install_packages}
\alias{install_packages}
\title{Install R Packages Locally}
\usage{
install_packages(packages, local_path = "~/local_R_lib")
}
\arguments{
\item{packages}{A character vector of package names to be installed.}

\item{local_path}{A string specifying the path to the local library where
packages should be installed. Defaults to "~/local_R_lib".}
}
\description{
The function `install_packages` checks for the presence of specified packages
and installs any that are missing into a local library path. The local library
path is specified by the user and defaults to "~/local_R_lib". If the specified
directory does not exist, the function will create it. After installing packages,
the local library path is added to the session's library paths, allowing
immediate use of the installed packages.

To configure the `local_path`, specify a valid directory path as a string.
This path will be used to store the installed packages. For example, setting
`local_path` to "~/my_r_packages" will install the packages in the
"my_r_packages" directory in the user's home folder.
}
\details{
Installs specified R packages in a local directory. This function is
particularly useful for users who do not have permission to install packages
system-wide or wish to manage packages in a self-contained environment.
}
\examples{
\dontrun{
  install_packages(...)
}

}
