% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchSVR.R
\name{main_grid_search}
\alias{main_grid_search}
\title{Main Grid Search Function}
\usage{
main_grid_search(
  data_env,
  bcv_folds = 5,
  processed_data,
  kernel_type = "linear",
  lags_column = c(8),
  initial_column_values = c(1),
  lags_response = NULL,
  initial_prediction_value = 1,
  is_test_mode = FALSE,
  is_silent_mode = TRUE,
  start_time = NULL,
  should_plot_response = FALSE,
  generate_response_predictions_cv = TRUE,
  basic_filter_check = TRUE,
  show_progress_bar = FALSE
)
}
\arguments{
\item{data_env}{Data environment object containing datasets, model
parameters, and other settings for the grid search.}

\item{bcv_folds}{Integer specifying the number of folds for cross-validation,
defaults to 5.}

\item{processed_data}{Dataframe of processed data set to be used in the
model.}

\item{kernel_type}{Character string indicating the type of kernel to use,
default is "linear".}

\item{lags_column}{Integer or vector specifying the number of lags for the
columns, defaults to c(8).}

\item{initial_column_values}{Numeric or vector providing initial values for
columns, defaults to c(1).}

\item{lags_response}{Integer or vector specifying the number of lags for the
response, defaults to NULL.}

\item{initial_prediction_value}{Numeric value indicating the initial value
for prediction, defaults to 1.}

\item{is_test_mode}{Logical flag indicating if the function is running in
test mode, defaults to FALSE.}

\item{is_silent_mode}{Logical flag indicating if the function should operate
in silent mode, suppressing messages and output, defaults to TRUE.}

\item{start_time}{POSIXct time object indicating the start time of the
function, defaults to NULL.}

\item{should_plot_response}{Logical flag indicating whether the response
variable should be plotted, defaults to FALSE.}

\item{generate_response_predictions_cv}{Logical flag indicating if response
predictions should be generated, defaults to TRUE.}

\item{basic_filter_check}{Logical flag indicating if basic filters should
be applied, defaults to TRUE.}

\item{show_progress_bar}{Disables progress bar. Defaults to FALSE.}
}
\value{
A list containing the results of the grid search, time taken for the
 execution, additional model parameters such as SVM tolerance and cache size,
 maximum average correlation of final candidates, and number of final candidates.
}
\description{
This function performs a comprehensive grid search over specified parameters
for a given model, supporting both linear and non-linear kernels. It is
designed to optimize models by testing various combinations of parameters
and evaluating their performance.
}
\examples{
\dontrun{
  # Assuming 'data_env_list', 'processed_data', and 'norm_vsvr_response'
  # are predefined
  # Example with default parameters
  results <- main_grid_search(
    data_env_list = data_env_list,
    processed_data = processed_data,
    norm_vsvr_response = norm_vsvr_response
  )
}
}
