% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalHelpers.R
\name{max_diff}
\alias{max_diff}
\title{Calculate the Maximum Difference Between Consecutive Points in a Vector}
\usage{
max_diff(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
The maximum absolute difference between consecutive points in the
 vector.
}
\description{
This function computes the maximum absolute difference between consecutive
points in a numeric vector. It is useful for identifying the largest change
between any two adjacent values in the vector.
}
\examples{
vec <- c(1, 2, 4, 7, 11)
max_diff_value <- max_diff(vec)
# `max_diff_value` will be the largest difference between adjacent values
# in `vec`.

}
