% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize_all_signals}
\alias{normalize_all_signals}
\title{Normalize All Columns in a Data Frame Using Min-Max Scaling}
\usage{
normalize_all_signals(df = NULL)
}
\arguments{
\item{df}{Data frame containing the columns to be normalized.}
}
\value{
Data frame with normalized columns added, retaining the original
 columns.
}
\description{
This function normalizes all columns in a data frame using Min-Max scaling.
The normalized values are added as new columns with "_norm" appended to the
original column names.
}
\examples{
df <- data.frame(Time = 1:10,
                 Signal1 = c(2, 4, 6, 8, 10, 5, 3, 7, 9, 1),
                 Signal2 = c(10, 5, 8, 3, 6, 9, 2, 7, 4, 1))
normalized_df <- normalize_all_signals(df)
# `normalized_df` will include 'Time_norm', 'Signal1_norm', and 'Signal2_norm'.

}
