% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize_signals_by_name}
\alias{normalize_signals_by_name}
\title{Normalize Selected Columns in a Data Frame Using Min-Max Scaling}
\usage{
normalize_signals_by_name(df, signal_names)
}
\arguments{
\item{df}{Data frame containing the columns to be normalized.}

\item{signal_names}{Character vector specifying which columns should be normalized.}
}
\value{
Data frame with normalized columns added.
}
\description{
Normalize Selected Columns in a Data Frame Using Min-Max Scaling
}
\examples{
df <- data.frame(Time = 1:10, Signal1 = c(2, 4, 6, 8, 10, 5, 3, 7, 9, 1),
 Signal2 = c(10, 5, 8, 3, 6, 9, 2, 7, 4, 1))
normalized_df <- normalize_signals_by_name(df, c("Signal1", "Signal2"))

}
