% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeBrainModelWithPSO.R
\name{optimize_brain_model_with_PSO}
\alias{optimize_brain_model_with_PSO}
\title{Optimize Brain Model Using PSO}
\usage{
optimize_brain_model_with_PSO(
  config,
  psoptim_config,
  data,
  model,
  multi = FALSE,
  signal_names,
  excluded_cols = NULL,
  predictors_names,
  params_lower_bounds,
  params_upper_bounds,
  params_initial_values,
  vsvr_response,
  silent = FALSE,
  plot_response = TRUE
)
}
\arguments{
\item{config}{A list of configuration options for the brain modeler.}

\item{psoptim_config}{A list of configuration options for psoptim.}

\item{data}{A data frame containing the dataset to be optimized.}

\item{model}{Character string specifying the type of model to be optimized.
Should be one of "FIR", "NFIR", "ARX", or "NARX".}

\item{multi}{(Optional) Logical indicating if the model should be multivariate.
If TRUE, optimizes with two input variables; if FALSE, optimizes with one input variable.
Defaults to FALSE.}

\item{signal_names}{Vector of names representing signal columns.}

\item{excluded_cols}{(Optional) Vector of column names that should be excluded from optimization.
Defaults to NULL.}

\item{predictors_names}{Vector of predictor names for the model.}

\item{params_lower_bounds}{Vector of lower bounds for model optimization parameters.}

\item{params_upper_bounds}{Vector of upper bounds for model optimization parameters.}

\item{params_initial_values}{Vector of initial values for model optimization parameters.
Length should be equal for all parameter vectors.}

\item{vsvr_response}{Name of the column representing the VSVR response.}

\item{silent}{(Optional) A logical indicating if the function should run silently. Default is TRUE.}

\item{plot_response}{(Optional) A logical to decide whether to plot the response signal. Defaults to TRUE.}
}
\value{
An object with results from the psoptim optimization.
}
\description{
Utilizing blocked k-fold cross-validation, the package
 conducts hyperparameter optimization through PSO. Models generated
 include FIR, NFIR, ARX, and NARX. These models can be univariate or multivariate.
}
\details{
This function offers a comprehensive toolkit for the analysis and
modeling of biological signal data specific to individual patients.
It facilitates the training of SVR models to represent and predict
cerebral autoregulation phenomena.


This function depends on:
\itemize{
  \item Internal package functions:
  \itemize{
    \item \code{\link{configure_pso_brain_modeler}}: Configure the Brain Modeler for PSO
    \item \code{\link{configure_psoptim_control}}: Configure the control parameters for psoptim
    \item \code{\link{configure_data_env}}: Configure the data environment for modeling
    \item \code{\link{validate_inputs_main}}: Validate Main Inputs
    \item \code{\link{pso_model}}: PSO Model Function
  }
  \item External package function:
  \itemize{
    \item \code{psoptim}
  }
}
}
\examples{

\dontrun{
 mydata <-   data.frame(
  Feature1 = rnorm(100),
  Feature2 = rnorm(100),
  Feature3 = rnorm(100),
  Feature4 = rnorm(100)
  )
 brain_modeler_config <- configure_pso_brain_modeler()
 psoptim_config <- configure_psoptim_control()

 optimize_brain_model_with_PSO(config = brain_modeler_config,
   psoptim_config = psoptim_config,
   data = mydata,
   model = "FIR",
   multi = FALSE,
   signal_names = c("Feature1", "Feature2"),
   excluded_cols = c("Feature3", "Feature4"),
   predictors_names = c("Feature1"),
   params_lower_bounds = c(1, 0.1, 1),
   params_upper_bounds = c(10, 0.5, 3),
   params_initial_values = c(NA, NA, NA),
   vsvr_response = "Feature2",
   silent = TRUE)
}

}
