% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchSVR.R
\name{params_config}
\alias{params_config}
\title{Configure Parameters for SVR}
\usage{
params_config(kernel_type = "linear", is_test_mode = FALSE)
}
\arguments{
\item{kernel_type}{The type of kernel to use in SVR. Valid options are
"linear" and others (e.g., "radial").}

\item{is_test_mode}{Logical flag indicating whether the function is in
test mode.}
}
\value{
A list containing the parameters 'nu', 'cost', and 'gamma'.
}
\description{
This function configures the parameters for Support Vector Regression (SVR)
based on the kernel type and test mode.
}
\examples{
  params_config(kernel_type = "linear", is_test_mode = FALSE)
}
