% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalHelpers.R
\name{plot_vsvr_response_signal}
\alias{plot_vsvr_response_signal}
\title{Plot VSVR Response Signal}
\usage{
plot_vsvr_response_signal(signal, points = 70)
}
\arguments{
\item{signal}{A numeric vector representing the VSVR response signal.}

\item{points}{(Optional) The number of points to plot along the x-axis.
Defaults to 70.}
}
\description{
This function plots the VSVR (Vector Support Vector Regression) response
signal over time or instances.
It creates a line plot for the given signal with specified points.
}
\details{
The function generates a line plot with 'Time/Instance' on the x-axis and
 'Response' on the y-axis.
It is designed to provide a visual representation of the VSVR response signal.
}
\examples{
response_signal <- runif(70, min = -1, max = 1)
plot_vsvr_response_signal(response_signal)

}
