% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{process_dataframe}
\alias{process_dataframe}
\title{Process a Data Frame to Exclude, Normalize, and Lag Signals}
\usage{
process_dataframe(
  df = data.frame(),
  excluded_cols = NULL,
  lags = NULL,
  signals,
  lagged_signals = NULL
)
}
\arguments{
\item{df}{A data frame containing the signals to be processed. Providing an
empty data frame is not practical for processing; hence, it's recommended
 to always provide this argument.}

\item{excluded_cols}{A character vector specifying the columns to be excluded.
Defaults to NULL, meaning no columns are excluded if not specified.}

\item{lags}{An integer specifying the number of lags to add for each signal.
Defaults to NULL, implying no lags are added unless specified.}

\item{signals}{A character vector specifying the signals to be normalized.
This argument is mandatory.}

\item{lagged_signals}{A character vector specifying the signals to be lagged.
If NULL, all signals will be lagged. Defaults to NULL. The function appends
a '_norm' suffix to the signal names if not already present, assuming
normalized signals follow this naming convention.}
}
\value{
A new data frame with specified columns excluded, selected columns
 normalized, and lagged versions of the signals added. Rows with resulting
 NA values from lags are not explicitly handled and should be managed in
 subsequent steps of analysis.
}
\description{
This function processes a given data frame to exclude specified columns,
normalize selected columns, and add lagged versions of the signals. It first
excludes specified columns, then normalizes the signals, and finally adds
lagged versions. Note: The exclusion of signals post-normalization might be
redundant and is subject to review in future updates.
}
\examples{
df <- data.frame(Time = 1:10, A = c(2, 4, 6, 8, 10, 5, 3, 7, 9, 1),
                 B = c(10, 5, 8, 3, 6, 9, 2, 7, 4, 1))
processed_df <- process_dataframe(df = df,
                                  excluded_cols = c("Time"),
                                  lags = 2,
                                  signals = c("A", "B"))
}
