% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelObjectives.R
\name{pso_arx}
\alias{pso_arx}
\title{PSO Objective Function for ARX Modeling}
\usage{
pso_arx(params)
}
\arguments{
\item{params}{A numeric vector with the optimization parameters.
Its length should be 4 (without gamma) or 5 (with gamma).}
}
\value{
A numeric value that represents the objective value for the given parameters.
}
\description{
This function takes a numeric vector of optimization parameters and returns
the objective value for ARX modeling using the specified parameters.
}
\examples{
\dontrun{
  # Assuming appropriate context:
  pso_arx(c(0.5, 0.7, 1, 2))
  pso_arx(c(0.5, 0.7, 0.2, 1, 2))
}
}
