% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelObjectives.R
\name{pso_fir}
\alias{pso_fir}
\title{PSO Objective Function for FIR Modeling}
\usage{
pso_fir(params)
}
\arguments{
\item{params}{A numeric vector with the optimization parameters.
Length should be either 3 (without gamma) or 4 (with gamma).}
}
\value{
A numeric value representing the objective value for the given parameters.
}
\description{
This function takes a numeric vector of optimization parameters and returns
the objective value for the FIR modeling using the specified parameters.
}
\examples{
# Assuming appropriate context:
pso_fir(c(0.5, 0.7, 1))
pso_fir(c(0.5, 0.7, 1, 0.2))
}
