% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{pso_objective_selector}
\alias{pso_objective_selector}
\title{PSO Objective Function Selector}
\usage{
pso_objective_selector(model, multi = FALSE)
}
\arguments{
\item{model}{A character string indicating the model to optimize. Should be one of "FIR", "NFIR", "ARX", "NARX".}

\item{multi}{A logical indicating if multi-objective optimization is being conducted.}
}
\value{
A function reference to the selected objective function.
}
\description{
This function selects an appropriate objective function based on the model provided.
}
\examples{
main_pso_objective("FIR")
}
