% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearchSVR.R
\name{svr_grid_search}
\alias{svr_grid_search}
\title{Perform a Grid Search for Support Vector Regression}
\usage{
svr_grid_search(
  config,
  dataset,
  kernel_type,
  is_multivariate = FALSE,
  signal_names,
  predictor_names,
  response_var,
  exclude_columns = NULL,
  lags_column = c(8),
  lags_response = NULL,
  initial_column_value = c(1),
  initial_prediction_value = 1,
  extra_column_name = NULL,
  should_plot_response = FALSE,
  is_test_mode = FALSE,
  is_silent_mode = TRUE,
  generate_response_predictions_cv = FALSE,
  basic_filter_check = TRUE,
  show_progress_bar = FALSE
)
}
\arguments{
\item{config}{List of configuration settings for the grid search.}

\item{dataset}{Dataframe containing the data to be used in the grid search.}

\item{kernel_type}{Character string specifying the type of kernel to be used
in SVR.}

\item{is_multivariate}{Logical flag indicating if a multivariate analysis is
to be performed.}

\item{signal_names}{Character vector naming the signal variables.}

\item{predictor_names}{Character vector naming the predictor variables.}

\item{response_var}{Character string naming the response variable for SVR.}

\item{exclude_columns}{Character vector specifying columns to be excluded
from the analysis. Defaults to NULL.}

\item{lags_column}{Integer or vector specifying the number of lags for the
columns. Defaults to c(8).}

\item{lags_response}{Integer or vector specifying the number of lags for the
response. Defaults to NULL.}

\item{initial_column_value}{Numeric value or vector providing initial values
for columns. Defaults to c(1).}

\item{initial_prediction_value}{Numeric value providing the initial value
for prediction. Defaults to 1.}

\item{extra_column_name}{Character string specifying the name of an extra
column to be included. Defaults to NULL.}

\item{should_plot_response}{Logical flag indicating if the response should
be plotted. Defaults to FALSE.}

\item{is_test_mode}{Logical flag indicating if the function is in test mode.
Defaults to FALSE.}

\item{is_silent_mode}{Logical flag indicating if the function should run in
silent mode, minimizing output. Defaults to TRUE.}

\item{generate_response_predictions_cv}{Logical flag indicating if response
predictions should be generated. Defaults to FALSE.}

\item{basic_filter_check}{Logical flag indicating if a basic filter check
should be applied. Defaults to TRUE.}

\item{show_progress_bar}{Disables progress bar. Defaults to FALSE.}
}
\value{
A list containing the results of the grid search, potentially
 including the best model parameters, performance metrics, and any generated response predictions.
}
\description{
This function performs a grid search to find the optimal parameters for a
Support Vector Regression (SVR) model. It allows for various configurations
and settings, facilitating extensive testing and optimization of the SVR model.
}
\examples{
\dontrun{
  svr_grid_search(my_config, my_data, "linear", ...)
}
}
