% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_character_vector}
\alias{validate_character_vector}
\title{Validate Character Vector}
\usage{
validate_character_vector(vec, name)
}
\arguments{
\item{vec}{A vector. The vector to validate.}

\item{name}{A character. A name or description for the vector, used in error messages.}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Ensures that the provided vector is a non-empty character vector.
}
\examples{
validate_character_vector(c("a", "b"), "Example")

}
