% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_character_vector_list}
\alias{validate_character_vector_list}
\title{Validate Character Vector List}
\usage{
validate_character_vector_list(char_list)
}
\arguments{
\item{char_list}{A list. The list that should contain character vectors.}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Ensures that the provided input is a non-empty list with character vector elements.
}
\examples{
validate_character_vector_list(list(c("a", "b"), c("x", "y")))

}
