% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data Frame}
\usage{
validate_data(data)
}
\arguments{
\item{data}{A data frame. The data to validate.}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Ensures that the provided data is a non-empty data frame.
}
\examples{
validate_data(data.frame(a=1:5, b=6:10))

}
