% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_data_partitions}
\alias{validate_data_partitions}
\title{Validate Data Partitions}
\usage{
validate_data_partitions(
  data_partitions,
  blocks = 5,
  partitions_names = c("training", "validation")
)
}
\arguments{
\item{data_partitions}{List. List containing the data partitions.}

\item{blocks}{Numeric (optional). Expected number of blocks (partitions) in the data_partitions. Default is 5.}

\item{partitions_names}{Character vector (optional). Expected names for the elements inside each partition. Defaults to c("training", "validation").}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Ensures that the provided data partitions are structured correctly.
}
\examples{
data_partitions <- list(list(training=data.frame(a=1:3), validation=data.frame(a=4:5)))
validate_data_partitions(data_partitions, 1)

}
