% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{validate_df_and_signals}
\alias{validate_df_and_signals}
\title{Validate Input Data Frame and Signal Names}
\usage{
validate_df_and_signals(df, signal_names, exclude_mode = FALSE)
}
\arguments{
\item{df}{A data frame to validate.}

\item{signal_names}{A character vector specifying the names of the columns to validate.}

\item{exclude_mode}{Logical. If TRUE, indicates validation for exclusion, else for inclusion.}
}
\value{
Invisible NULL if validation passes. Stops with an error if there are issues.
}
\description{
This function validates the provided data frame and the specified column (signal) names.
}
