% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{validate_inputs_main}
\alias{validate_inputs_main}
\title{Validate Main Inputs}
\usage{
validate_inputs_main(
  data,
  model,
  signal_names,
  predictors_names,
  vsvr_response,
  excluded_cols = NULL,
  multi = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data}{A dataframe with the data.}

\item{model}{A character string specifying the model. Valid values are "FIR", "NFIR", "ARX", and "NARX".}

\item{signal_names}{A character vector with signal names.}

\item{predictors_names}{A character vector with predictor names.}

\item{vsvr_response}{A single character string indicating the response variable for v-SVR.}

\item{excluded_cols}{(Optional) A character vector with column names to be excluded. Defaults to NULL.}

\item{multi}{(Optional) A logical indicating whether multi-response is used. Default is FALSE.}

\item{silent}{(Optional) A logical indicating if the function should run silently. Default is TRUE.}
}
\description{
This function validates the main inputs for data processing and model building.
}
\examples{
data <- data.frame(signal = rnorm(100), predictor = rnorm(100))
PSOBrainModeler:::validate_inputs_main(data = data,
  model = "FIR",
  signal_names = c("signal"),
  predictors_names = c("predictor"),
  vsvr_response = "signal")

}
