% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_logical}
\alias{validate_logical}
\title{Validate Logical Value}
\usage{
validate_logical(log_val, name)
}
\arguments{
\item{log_val}{Logical. The logical value to validate.}

\item{name}{Character. A name or description for the logical value, used in error messages.}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Ensures that the provided value is a logical scalar.
}
\examples{
validate_logical(TRUE, "Example")

}
