% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_model}
\alias{validate_model}
\title{Validate Model Type}
\usage{
validate_model(model)
}
\arguments{
\item{model}{Character. The model type name.}
}
\value{
No return value, this function stops with a message if validation fails.
}
\description{
Checks if the provided model type is one of the expected model types.
}
\examples{
validate_model("FIR")

}
