% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{validate_pso_svr_params}
\alias{validate_pso_svr_params}
\title{Validate Parameters for PSO-SVR}
\usage{
validate_pso_svr_params(list)
}
\arguments{
\item{list}{A list containing the parameters to be validated. The list can
include 'cost', 'nu', 'gamma', and 'lags'.}
}
\value{
Invisible NULL. If any of the parameters fail validation, the function
        stops with an error message.
}
\description{
This function validates the parameters for the Particle Swarm Optimization
Support Vector Regression (PSO-SVR) model. It checks if the provided
parameters 'cost', 'nu', 'gamma', and 'lags' meet the expected criteria.
}
\details{
The function expects the following parameters in the list:
- 'cost': A positive numeric value.
- 'nu': A numeric value between 0 (exclusive) and 1 (inclusive).
- 'gamma': (Optional) A positive numeric value.
- 'lags': A numeric vector. The function validates only the first value, which
          should be a positive integer.
}
\examples{
parameters <- list(cost = 1.5, nu = 0.5, gamma = 0.01, lags = c(2, 3, 4))
PSOBrainModeler:::validate_pso_svr_params(parameters)

}
