% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSVMRegression.R
\name{vsvr_model}
\alias{vsvr_model}
\title{v-Support Vector Regression (v-SVR) Model}
\usage{
vsvr_model(data, response_var, cost, nu, gamma = NULL, tolerance)
}
\arguments{
\item{data}{Dataframe. Contains predictor variables and the response variable.}

\item{response_var}{Character. Name of the response variable in the data frame.}

\item{cost}{Numeric. Cost parameter for the SVR model.}

\item{nu}{Numeric. Nu parameter for the SVR model.}

\item{gamma}{Numeric (optional). Kernel parameter for the SVR model. Required for radial kernels.}

\item{tolerance}{Numeric. Tolerance for stopping criterion.}
}
\value{
An SVR model object from `e1071::svm`.
}
\description{
Fits a Nu Support Vector Regression (SVR) model using the `svm` function from the `e1071` package.
}
\examples{
# Fit an SVR model with a radial kernel
# model <- vsvr_model(data = my_data,
#                     response_var = "target",
#                     cost = 1,
#                     nu = 0.3,
#                     gamma = 0.01,
#                     tolerance = 0.001)
}
