% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{add_pressure_step}
\alias{add_pressure_step}
\title{Generate and Add a Smoothed Pressure Step to a Signal}
\usage{
add_pressure_step(
  pressure_start = 3L,
  signal_end = 40L,
  butter_order = 2L,
  butter_fs = 0.2
)
}
\arguments{
\item{pressure_start}{(Optional) Integer indicating the index at which the pressure step starts. Defaults to 3L.}

\item{signal_end}{(Optional) Integer indicating the index at which the pressure step ends. Defaults to 40L.}

\item{butter_order}{(Optional) Integer representing the order of the Butterworth filter. Defaults to 2L.}

\item{butter_fs}{(Optional) Numeric value representing the sampling frequency for the Butterworth filter. Defaults to 0.2.}
}
\value{
A data frame containing a single column, `signal`, with the smoothed pressure step.
}
\description{
This function creates a smoothed pressure step using a Butterworth filter
and returns the result as a data frame. The output signal starts with a zero value
up to `pressure_start` and then steps down to -1, smoothed using the Butterworth filter.
}
\examples{
pressure_step <- add_pressure_step(pressure_start = 10L, signal_end = 100L,
                                   butter_order = 2L, butter_fs = 0.2)

}
