% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{add_pressure_step}
\alias{add_pressure_step}
\title{Generate and Add a Smoothed Pressure Step to a Signal}
\usage{
add_pressure_step(
  pressure_start = 5,
  signal_end = 40,
  butter_order = 2,
  butter_fs = 0.2
)
}
\arguments{
\item{pressure_start}{Integer indicating the index at which the pressure step starts.}

\item{signal_end}{Integer indicating the index at which the pressure step ends.}

\item{butter_order}{Integer representing the order of the Butterworth filter.}

\item{butter_fs}{Numeric value representing the sampling frequency for the Butterworth filter.}
}
\value{
A data frame containing a single column, `signal`, with the smoothed pressure step.
}
\description{
This function creates a smoothed pressure step using a Butterworth filter
and returns the result as a data frame. The output signal starts with a zero value
up to `pressure_start` and then steps down to -1, smoothed using the Butterworth filter.
}
\examples{
pressure_step <- add_pressure_step(pressure_start = 10, signal_end = 100,
                                   butter_order = 2, butter_fs = 0.2)
plot(pressure_step$signal)

}
