% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockedCrossValidation.R
\name{blocked_cv}
\alias{blocked_cv}
\title{Generate Data Partitions for Blocked K-fold Cross-Validation}
\usage{
blocked_cv(data = NULL, num_blocks = 5, validation_size = 0.2)
}
\arguments{
\item{data}{Dataframe. The dataset to partition. Defaults to NULL.}

\item{num_blocks}{(Optional) Integer. The number of blocks to divide the dataset into. Defaults to 5.}

\item{validation_size}{(Optional) Numeric (between 0 and 1). Proportion of data to be used
for validation in each partition. Defaults to 0.2.}
}
\value{
A list of partitions, each containing a training set and a validation set.
}
\description{
Generate Data Partitions for Blocked K-fold Cross-Validation
}
\examples{
data <- data.frame(x = 1:100, y = 101:200)
partitions <- blocked_cv(data, num_blocks = 5, validation_size = 0.2)
}
