% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{configure_data_env}
\alias{configure_data_env}
\title{Configure the data environment for modeling}
\usage{
configure_data_env(
  config,
  data,
  excluded_cols,
  signal_names,
  predictors_names,
  vsvr_response
)
}
\arguments{
\item{config}{A list of configuration options.}

\item{data}{A data frame containing the dataset to be processed.}

\item{excluded_cols}{Vector of column names that should be excluded from processing.}

\item{signal_names}{Vector of names representing signal columns.}

\item{predictors_names}{Vector of predictor names for the model.}

\item{vsvr_response}{Name of the column representing the VSVR response.}
}
\value{
NULL. The function assigns values to global variables in the package environment.
}
\description{
This function prepares the data environment required for modeling.
}
