% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurations.R
\name{configure_data_env}
\alias{configure_data_env}
\title{Configure the data environment for modeling}
\usage{
configure_data_env(
  config,
  data,
  excluded_cols,
  signal_names,
  predictors_names,
  vsvr_response
)
}
\arguments{
\item{config}{A list of configuration options.}

\item{data}{A data frame containing the dataset to be processed.}

\item{excluded_cols}{Vector of column names that should be excluded from processing.}

\item{signal_names}{Vector of names representing signal columns. The resulting normalized signal names
in the processed data will have a "_norm" suffix.}

\item{predictors_names}{Vector of predictor names for the model. The resulting normalized predictor names
in the processed data will have a "_norm" suffix.}

\item{vsvr_response}{Name of the column representing the VSVR response. The resulting normalized name
in the processed data will have a "_norm" suffix.}
}
\value{
A list containing various data configurations including the processed data with normalized signal names.
}
\description{
This function prepares the data environment required for modeling. The signals in the resulting data
are normalized and the signal names have a "_norm" suffix to indicate this normalization.
}
\details{
This function depends on:
\itemize{
  \item Internal package functions:
  \itemize{
    \item \code{\link{add_pressure_step}}: Generate a dataframe for the smoothed negative
   pressure step response with a Butterworth filter.
    \item \code{\link{process_dataframe}}: Exclude, Normalize, and Lag Signals of Data Frame.
The normalization will append a "_norm" suffix to the signal names.
    \item \code{\link{blocked_cv}}: Perform k-folded blocked cross-validation.
  }
}
}
\examples{
config <- PSOBrainModeler:::configure_pso_brain_modeler()
data <- data.frame(feature1 = rnorm(10), feature2 = rnorm(10))
PSOBrainModeler:::configure_data_env(config = config, data = data, excluded_cols = NULL,
  signal_names = c("feature1", "feature2"), predictors_names = c("feature1"),
  vsvr_response = "feature2")

}
