% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockedCrossValidation.R
\name{cross_validate_partition}
\alias{cross_validate_partition}
\title{Cross-Validate Partition for SVR Model}
\usage{
cross_validate_partition(
  cost,
  nu,
  gamma,
  data_partitions,
  bcv_folds,
  signal_norm_names,
  predictors_norm_names,
  lagged_cols,
  col_lags,
  vsvr_response,
  vsvr_tolerance
)
}
\arguments{
\item{cost}{Numeric. The cost parameter for the SVR model.}

\item{nu}{Numeric. The nu parameter for the SVR model.}

\item{gamma}{Numeric. The gamma parameter for the SVR model.}

\item{data_partitions}{List of dataframes. The data partitions to use for training and validation.}

\item{bcv_folds}{Numeric. Number of cross-validation folds.}

\item{signal_norm_names}{Character vector. Normalized signal column names.}

\item{predictors_norm_names}{Character vector. Normalized predictor column names.}

\item{lagged_cols}{Character vector. Names of lagged columns.}

\item{col_lags}{Numeric vector. Number of lags for each column.}

\item{vsvr_response}{Character. Response column name for the SVR model.}

\item{vsvr_tolerance}{Numeric. Tolerance parameter for the SVR model.}
}
\value{
A list containing the average correlation (`avg_cor`) and average mean squared error (`avg_error`).
}
\description{
This function performs cross-validation on the given data partitions using a Support Vector Regression (SVR) model.
It returns the average correlation and mean squared error over the partitions.
}
\examples{

data_partition_sample <- list(list(training = data.frame(feature1 = rnorm(20),
feature2 = rnorm(20), feature1_1 = rnorm(20)), validation = data.frame(
feature1 = rnorm(20), feature2 = rnorm(20), feature1_1 = rnorm(20))))
cross_validate_partition(cost = 1, nu = 0.5, gamma = NULL, data_partition_sample,
bcv_folds = 1, signal_norm_names = c("feature1", "feature2"),
predictors_norm_names = c("feature1"), lagged_cols = c("feature1"), col_lags = c(1),
vsvr_response = "feature2", vsvr_tolerance = 1)

}
