% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOHelpers.R
\name{extract_and_round_pso_params}
\alias{extract_and_round_pso_params}
\title{Extract and Round Parameters from PSO Optimization}
\usage{
extract_and_round_pso_params(params, has_gamma = FALSE, n_lags = 1)
}
\arguments{
\item{params}{A numeric vector with the optimization parameters.}

\item{has_gamma}{A logical indicating if the gamma parameter is included.}

\item{n_lags}{An integer for the number of lags being optimized.}
}
\value{
A list containing the rounded parameters including cost, nu, gamma (if applicable), and lags.
}
\description{
This function takes a numeric vector of parameters, rounds them as necessary,
and structures them in a list for easier use in subsequent processes.
}
\examples{
params <- c(1.234, 0.5678, 0.00089, 2.345)
extract_and_round_pso_params(params, has_gamma = TRUE, n_lags = 1)
}
