% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{generate_signal_response_predictions_helper}
\alias{generate_signal_response_predictions_helper}
\title{Helper function to generate signal response predictions}
\usage{
generate_signal_response_predictions_helper(
  data,
  col_lags,
  response_lags,
  initial_column_values,
  prediction_initial_value,
  cost,
  nu,
  gamma,
  data_list
)
}
\arguments{
\item{data}{Data frame containing the dataset to be processed.}

\item{col_lags}{Numeric vector indicating the lags for the columns.}

\item{response_lags}{Numeric vector indicating the response lags (default: NULL).}

\item{initial_column_values}{Numeric vector indicating initial values for columns.}

\item{prediction_initial_value}{Numeric value indicating the initial prediction value.}

\item{cost}{A numeric value for the cost parameter.}

\item{nu}{A numeric value for the nu parameter.}

\item{gamma}{A numeric value for the gamma parameter (default: NULL).}

\item{data_list}{A list containing various data configurations}
}
\value{
Result from the generate_signal_response_predictions function.
}
\description{
Constructs parameters for the generate_signal_response_predictions function. This
involves predicting signal responses based on the provided parameters and data.
}
