% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{generate_time_series_data}
\alias{generate_time_series_data}
\title{Generate Time-series Model Data}
\usage{
generate_time_series_data(
  input_df,
  data_cols,
  predictor_cols,
  lagged_cols,
  lag_values,
  is_training,
  vsvr_response
)
}
\arguments{
\item{input_df}{Dataframe containing the raw data.}

\item{data_cols}{Character vector of column names to include in the training
set.}

\item{predictor_cols}{Character vector of column names to include in the
prediction set.}

\item{lagged_cols}{Character vector specifying which columns should have
lagged values.}

\item{lag_values}{Integer vector specifying the number of lags for each
lagged column.
                 Should be in the same order as 'lagged_cols'.}

\item{is_training}{Boolean indicating whether the data is for training (TRUE)
or prediction (FALSE).}

\item{vsvr_response}{Character representing the response column name for the
SVR model.}
}
\value{
Dataframe containing the specified columns and their lagged values,
 adjusted for training or prediction.
}
\description{
This function constructs a dataset tailored for training or predicting with
a time-series model.
It selects specified columns and incorporates their lagged values based on
the training or prediction requirement.
}
\examples{
\dontrun{
  generate_time_series_data(...)
}

}
