% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lagging.R
\name{lag_normalized_signals}
\alias{lag_normalized_signals}
\title{Add Lagged Columns to a Normalized Data Frame}
\usage{
lag_normalized_signals(data_frame = NULL, lags = 0, df_col_names = NULL)
}
\arguments{
\item{data_frame}{Data frame to which lagged columns will be added.}

\item{lags}{Integer specifying the number of lags for each column.}

\item{df_col_names}{Character vector containing names of columns to be lagged.}
}
\value{
A modified data frame with lagged columns and rows with NA values removed.
}
\description{
This function adds lagged columns for specified normalized columns in a data frame.
The columns to be lagged should be already normalized and have "_norm" at the end of their names.
}
\examples{
df <- data.frame(A_norm = c(0, 0.2, 0.4, 0.6, 0.8, 1))
lagged_df <- lag_normalized_signals(df, 2, c("A_norm"))
print(lagged_df)

}
