% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize_all_signals}
\alias{normalize_all_signals}
\title{Normalize All Columns in a Data Frame Using Min-Max Scaling}
\usage{
normalize_all_signals(df = NULL)
}
\arguments{
\item{df}{Data frame containing the columns to be normalized.}
}
\value{
Data frame with normalized columns added.
}
\description{
Normalize All Columns in a Data Frame Using Min-Max Scaling
}
\examples{
df <- data.frame(Time = 1:10, Signal1 = c(2, 4, 6, 8, 10, 5, 3, 7, 9, 1),
 Signal2 = c(10, 5, 8, 3, 6, 9, 2, 7, 4, 1))
normalized_df <- normalize_all_signals(df)

}
