% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeBrainModelWithPSO.R
\name{optimize_brain_model_with_PSO}
\alias{optimize_brain_model_with_PSO}
\title{Optimize Brain Model Using PSO}
\usage{
optimize_brain_model_with_PSO(
  config,
  psoptim_config,
  data,
  model,
  multi,
  signal_names,
  excluded_cols,
  predictors_names,
  params_lower_bounds,
  params_upper_bounds,
  params_initial_values,
  vsvr_response
)
}
\arguments{
\item{config}{A list of configuration options for the brain modeler.}

\item{psoptim_config}{A list of configuration options for psoptim.}

\item{data}{A data frame containing the dataset to be optimized.}

\item{model}{Type of model to be optimized (e.g., FIR, NFIR).}

\item{multi}{Logical indicating if the model should be multivariate.}

\item{signal_names}{Vector of names representing signal columns.}

\item{excluded_cols}{Vector of column names that should be excluded from optimization.}

\item{predictors_names}{Vector of predictor names for the model.}

\item{params_lower_bounds}{Vector of lower bounds for model parameters.}

\item{params_upper_bounds}{Vector of upper bounds for model parameters.}

\item{params_initial_values}{Vector of initial values for model parameters.}

\item{vsvr_response}{Name of the column representing the VSVR response.}
}
\value{
An object with results from the psoptim optimization.
}
\description{
Utilizing blocked k-fold cross-validation, the package
conducts hyperparameter optimization through PSO. Models generated
include FIR, NFIR, ARX, and NARX. These models can be univariate or multivariate.
}
\details{
This function offers a comprehensive toolkit for the analysis and
modeling of biological signal data specific to individual patients.
It facilitates the training of SVR models to represent and predict
cerebral autoregulation phenomena.
}
