% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{process_dataframe}
\alias{process_dataframe}
\title{Process a Data Frame to Exclude, Normalize, and Lag Signals}
\usage{
process_dataframe(
  df = data.frame(),
  excluded_cols,
  lags = NULL,
  signals,
  lagged_signals = NULL
)
}
\arguments{
\item{df}{A data frame containing the signals to be processed.}

\item{excluded_cols}{A character vector specifying the columns to be excluded.}

\item{lags}{An integer specifying the number of lags to add for each signal.}

\item{signals}{A character vector specifying the signals to be excluded after normalization.}

\item{lagged_signals}{A character vector specifying the signals to be lagged.
If NULL, all signals will be lagged.}
}
\value{
A new data frame with specified columns excluded, all columns normalized,
        and lagged versions of the signals added. Rows with resulting NA values
         from lags are omitted.
}
\description{
This function processes a given data frame to exclude specified columns,
normalize all columns, and add lagged versions of the signals.
}
\examples{
df <- data.frame(Time = 1:10, A = c(2, 4, 6, 8, 10, 5, 3, 7, 9, 1),
 B = c(10, 5, 8, 3, 6, 9, 2, 7, 4, 1))
processed_df <- process_dataframe(df, excluded_cols = c("Time"), lags = 2, signals = c("A", "B"))

}
