% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelObjectives.R
\name{pso_model}
\alias{pso_model}
\title{PSO Model Function}
\usage{
pso_model(params, model, multi = FALSE)
}
\arguments{
\item{params}{A numeric vector of parameters.}

\item{model}{A character string specifying the model. Valid values are "FIR", "NFIR", "ARX", and "NARX".}

\item{multi}{A logical indicating whether multi-response is used.}
}
\value{
The result of the `pso_training_model`.
}
\description{
This function performs the Particle Swarm Optimization (PSO) model based on the provided parameters and model.
}
