% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelObjectives.R
\name{pso_model}
\alias{pso_model}
\title{Particle Swarm Optimization (PSO) Model Function}
\usage{
pso_model(
  params,
  max_function_count = 1000,
  multi = FALSE,
  data_list,
  model,
  silent = TRUE,
  plot_response = TRUE,
  initial_response_value = 1,
  model_parameters,
  bcv_folds = 5,
  pso_env,
  seed = 123,
  progress_bar = NULL,
  generate_response_predictions_cv = FALSE,
  basic_filter_check = TRUE,
  fn_count_threshold = 30,
  fitness_accuracy = 3,
  round_accuracy = 2,
  signif_accuracy = 3,
  show_progress_bar = FALSE,
  minimum_candidates = 10,
  fn_start_threshold = 100,
  cv_folds_ratio = 0.2,
  time_on_fitness = FALSE,
  penalization_weight = 0.5
)
}
\arguments{
\item{params}{Parameter list for the PSO model.}

\item{max_function_count}{Maximum number of function evaluations. Defaults
to 1000.}

\item{multi}{Boolean flag to indicate if the model is multivariate. Defaults
to FALSE.}

\item{data_list}{A list containing processed data relevant to the model.}

\item{model}{The model type to be optimized.}

\item{silent}{Flag to control the display of messages during execution.
Defaults to TRUE.}

\item{plot_response}{Flag to enable or disable plotting the response.
Defaults to TRUE.}

\item{initial_response_value}{Initial value for the response variable.
Defaults to 1.}

\item{model_parameters}{List of specific model parameters.}

\item{bcv_folds}{Number of folds for blocked cross-validation. Defaults to 5.}

\item{pso_env}{PSO environment configuration.}

\item{seed}{Seed for random number generation. Defaults to 123.}

\item{progress_bar}{Progress bar configuration for the optimization process.}

\item{generate_response_predictions_cv}{Flag to generate response
predictions. Defaults to FALSE.}

\item{basic_filter_check}{Flag to enable basic filtering of the data.
Defaults to TRUE.}

\item{fn_count_threshold}{Threshold for function count in optimization.
Defaults to 30.}

\item{fitness_accuracy}{Numeric value specifying fitness evaluation accuracy; default is 3.}

\item{round_accuracy}{Numeric value for rounding off the parameters.
Specifies the number of decimal places for rounding.}

\item{signif_accuracy}{Numeric value for significant figure accuracy.
Defines the number of significant digits to retain.}

\item{show_progress_bar}{Disables progress bar. Defaults to FALSE.}

\item{minimum_candidates}{Sets the lower limit for candidate consideration.}

\item{fn_start_threshold}{Determines the function's starting point if there are no candidates.}

\item{cv_folds_ratio}{Specifies the proportion of data used for cross-validation.}

\item{time_on_fitness}{Apply time to the objective function.}

\item{penalization_weight}{Numeric value for the weight in optimization penalization.}
}
\value{
The result of the PSO training model function, including any metrics,
        model parameters, and performance indicators.
}
\description{
Executes PSO modeling using provided parameters and configurations.
This function is tailored for both univariate and multivariate models and
offers options for silent operation, response plotting, and integrates
various model parameters and PSO environment settings.
}
\examples{
\dontrun{
  pso_model(params, data_list = my_data_list, model = "my_model")
}

}
