% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelOptimization.R
\name{pso_training_model}
\alias{pso_training_model}
\title{Model training function using PSO optimization}
\usage{
pso_training_model(
  cost,
  nu,
  gamma = NULL,
  col_lags,
  response_lags = NULL,
  vsvr_response = NULL
)
}
\arguments{
\item{cost}{A numeric value for the cost parameter.}

\item{nu}{A numeric value for the nu parameter.}

\item{gamma}{A numeric value for the gamma parameter (default: NULL).}

\item{col_lags}{Numeric vector indicating the lags for the columns.}

\item{response_lags}{Numeric vector indicating the response lags (default: NULL).}

\item{vsvr_response}{Name of the column representing the VSVR response. (default: NULL).}
}
\value{
A numeric value that represents the optimization result for minimization.
}
\description{
This function handles the training process for a model based on given parameters
using Particle Swarm Optimization (PSO). It performs cross-validation, response
predictions, and signal evaluation.
}
