% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOModelOptimization.R
\name{pso_training_model}
\alias{pso_training_model}
\title{Particle Swarm Optimization (PSO) Training Model Function}
\usage{
pso_training_model(
  max_function_count = 1000,
  cost,
  nu,
  gamma = NULL,
  col_lags,
  response_lags = NULL,
  data_list,
  silent = TRUE,
  plot_response = FALSE,
  initial_column_values = c(1),
  prediction_initial_value = 1,
  bcv_folds = 5,
  pso_env,
  seed = 123,
  progress_bar = NULL,
  generate_response_predictions_cv = TRUE,
  basic_filter_check = TRUE,
  fn_count_threshold = 30,
  fitness_accuracy = 3,
  show_progress_bar = FALSE,
  minimum_candidates = 10,
  fn_start_threshold = 100,
  cv_folds_ratio = 0.2,
  time_on_fitness = FALSE,
  penalization_weight = 0.5
)
}
\arguments{
\item{max_function_count}{Maximum number of function evaluations, defaults
to 1000.}

\item{cost}{Cost parameter for the model.}

\item{nu}{Nu parameter for the model.}

\item{gamma}{Gamma parameter for the model, if applicable.}

\item{col_lags}{Column lag values for the model.}

\item{response_lags}{Response lag values for the model, if applicable.}

\item{data_list}{List containing the data required for the model.}

\item{silent}{Flag to control verbosity during the training process.}

\item{plot_response}{Flag to enable or disable plotting the response.}

\item{initial_column_values}{Initial column values for the model.}

\item{prediction_initial_value}{Initial prediction value for the model.}

\item{bcv_folds}{Number of folds for blocked cross-validation.}

\item{pso_env}{PSO environment settings.}

\item{seed}{Seed for random number generation.}

\item{progress_bar}{Progress bar for visualization of the optimization
process.}

\item{generate_response_predictions_cv}{Flag to generate response predictions.}

\item{basic_filter_check}{Logical flag to enable basic filtering of the data.}

\item{fn_count_threshold}{Integer setting the function count threshold for optimization.}

\item{fitness_accuracy}{Numeric value specifying fitness evaluation accuracy.}

\item{show_progress_bar}{Disables the progress bar display.}

\item{minimum_candidates}{Sets the lower limit for candidate consideration.}

\item{fn_start_threshold}{Determines the function's starting point if there are no candidates.}

\item{cv_folds_ratio}{Specifies the proportion of data used for cross-validation.}

\item{time_on_fitness}{Apply time to the objective function.}

\item{penalization_weight}{Numeric value for the weight in optimization penalization.}
}
\value{
Depending on the stage and result of the optimization, it can return
 different values, including early termination or the current optimization
 score.
}
\description{
Conducts PSO training with a given set of parameters, data, and environmental
settings.
It includes functionality for progress tracking, data processing, response
prediction generation,
signal scoring, and optimization. This function is designed to handle various
aspects of PSO training,
including parameter tuning, cross-validation, and response evaluation.
}
\examples{
\dontrun{
result <- pso_training_model(cost = 0.5, nu = 0.1, col_lags = c(3),
                            data_list = my_data_list, pso_env = my_pso_env,
                            progress_bar = my_progress_bar)
}

}
